Strongly type defaultFormFields
