+++ packages/ui/src/helpers/authenticator/formFields/index.ts
@@ -1,2 +1,3 @@
 export * from './formFields';
+export * from './defaults';
 export * from './util';
